CEDVA/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
PyMySQL==1.0.2
sqlparse==0.4.2
tzdata==2022.2
